Rename `get token` to `get account data`
The names of action `GET_IMGUR_TOKEN` or saga `getImgurToken` are actually confusing because not only a token is returned in the response.
To be able to send requests to Imgur, the token and the username are required, both are returned in the response, so `account data` is a better name.