Add saga for getting the images of an album
To get all images of an album another request to Imgur is necessary.
The response contains all data of the requested album.
Those data is already stored in the state, with the exception of the images.
For adding the images data to a certain album, only the id and the images need to be returned in the payload.
The images are indexed by their ids to find them easily in the Redux state.